[PATCH] Makefile.am: make sure libintl is in LDFLAGS when needed
The build-system internally builds a couple of static libraries used to
build the final programs.

When linking these libraries and the final programs, we need to
explicitly tell libtool to link with libintl in order to resolve all the
missing symbols.

This is done in the appropriate LDADD/LIBADD variables (or the per-target
*_LDADD/*_LIBADD ones when they do not default to LDADD or LIBADD), for
any sources using gettext/libintl functionalities.

This change fixes build failures [2,3] in cross-compilation, triggered by
Buildroot buildfarms [1].

[1] http://autobuild.buildroot.org/
[2] http://autobuild.buildroot.org/?reason=tpm-tools-1.3.8
[3] http://autobuild.buildroot.org/results/c82/c82a6c1f41c561b15f10e7b7ffcaed15fa0ec895/build-end.log

	Signed-off-by: Samuel Martin <s.martin49@gmail.com>
	Signed-off-by: Hon Ching (Vicky) Lo <honclo@linux.vnet.ibm.com>